Lazy solution
A grid size is assumed (change to your value in the file) and each square is populated with the color closest in the input. This doesn't solve all the technical issues with grid detection and such, it just provides a quick result that works, with some manual work. Cut the frame out of the QR itself before application, and avoid tilted images.